Update wp-approve-user.pot
Clarify and simplify English language text in interface. "Deny" and "Reject" are more correct and simpler than "Disapprove." It's not an "error" when someone tries to log in who is not yet approved or denied. "Log in" is the correct form when referring to the action of logging in where "log" is a verb. "Login" is used as a noun or adjective. https://www.grammar.com/log_in_vs._login